+++ CHANGELOG.md
@@ -6,7 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 2021-01-09 - Edu
-## changed
+
+### Added
+- **piped_processes.c** prepare info needed for all process 
+- **utils/lst_process_handler.c**  Include basic functions to handle process list (new, add_front, add_back, free_list & print)
+- **utils/lst_redir_handler.c**  Include basic functions to handle redir list (new, add_front, add_back, free_list & print)
+### Changed
+- **eval_token** just return token type, not path
+
+
+## 2021-01-09 - Edu
+## Changed
 - **utils/error_handler.c** replace usleep with bash sleep through execve 
 
 ## 2021-01-05 - Edu